add: 加入订单列表

+++ app/src/main/java/com/nhbs/fenxiao/module/store/bean/OrderManagerInfoBean.kt
@@ -30,7 +30,7 @@ data class OrderInfo(
     var title: String,
     var totalPrice: String,
     var type: Int,
-    var updateTime: Int,
+    var updateTime: Long,
     var userId: String,
     var userName: String
 ) : Serializable
